+++ src/ts/magma/api/collect/RangeHead.ts
@@ -1,8 +1,7 @@
+import { None } from "../../../magma/api/option/None";
 import { Option } from "../../../magma/api/option/Option";
-import { Main } from "../../../magma/app/Main";
+import { Some } from "../../../magma/api/option/Some";
 import { Head } from "./Head";
-import { Main.None } from "./Main.None";
-import { Main.Some } from "./Main.Some";
 export class RangeHead implements Head<number> {
 	length: number;
 	counter: number;
@@ -12,10 +11,10 @@ export class RangeHead implements Head<number> {
 	}
 	next(): Option<number> {
 		if (this.counter >= this.length){
-			return new Main.None<number>();
+			return new None<number>();
 		}
 		let value = this.counter;
 		this.counter++;
-		return new Main.Some<number>(value);
+		return new Some<number>(value);
 	}
 }
