+++ src/ts/magma/jvm/io/Files.ts
@@ -7,7 +7,9 @@ import { HeadedQuery } from "../../../magma/api/collect/query/HeadedQuery";
 import { Query } from "../../../magma/api/collect/query/Query";
 import { IOError } from "../../../magma/api/io/IOError";
 import { Path } from "../../../magma/api/io/Path";
+import { None } from "../../../magma/api/option/None";
 import { Option } from "../../../magma/api/option/Option";
+import { Some } from "../../../magma/api/option/Some";
 import { Result } from "../../../magma/api/result/Result";
 import { Main } from "../../../magma/app/Main";
 import { IOException } from "../../../java/io/IOException";
